requirements.txt
aiogram<3
uvloop
ujson
python-dotenv
aiohttp_socks